requirements.txt
numpy==1.24.3
pandas==2.0.3
xgboost==1.7.6
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
shap==0.42.1
pyyaml==6.0.1
streamlit==1.28.0
openpyxl==3.1.2
jupyter==1.0.0
